requirements.txt
crewai
crewai-tools
pydantic-ai
pydantic-ai-slim[groq]
python-dotenv
requests
beautifulsoup4
litellm